Add NeoForge stub for build
