+++ common/build.gradle
@@ -1,0 +1,13 @@
+plugins {
+    id 'fabric-loom' version '1.11-SNAPSHOT'
+}
+
+base {
+    archivesName = project.archives_base_name + "-common"
+}
+
+dependencies {
+    minecraft "com.mojang:minecraft:${minecraft_version}"
+    mappings "net.fabricmc:yarn:${yarn_mappings}:v2"
+    modImplementation "net.fabricmc:fabric-loader:${loader_version}"
+}
