+++ common/src/main/java/org/tecna/easypets/IndyPetsHelper.java
