+++ common/src/main/java/org/tecna/easypets/mixin/PetNaturalRegenMixin.java
