+++ common/src/main/java/org/tecna/easypets/mixin/ServerPlayerEntityMixin.java
