+++ common/src/main/resources/easypets.mixins.json
