+++ gradle.properties
@@ -12,3 +12,5 @@ archives_base_name=followersloadchunks
 # Dependencies
 # check this on https://modmuss50.me/fabric.html
 fabric_version=0.129.0+1.21.7
+neoforge_version=21.0.110-beta
+paper_version=1.20.4-R0.1-SNAPSHOT
