+++ neoforge/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    implementation project(":common")
+    implementation "net.neoforged:neoforge:${neoforge_version}:userdev"
+}
