+++ neoforge/src/main/java/net/neoforged/fml/common/Mod.java
@@ -1,0 +1,12 @@
+package net.neoforged.fml.common;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface Mod {
+    String value();
+}
