+++ neoforge/src/main/java/org/tecna/easypets/neoforge/EasyPetsNeoForge.java
@@ -1,0 +1,14 @@
+package org.tecna.easypets.neoforge;
+
+import net.neoforged.fml.common.Mod;
+import org.tecna.easypets.EasyPets;
+
+/**
+ * NeoForge entry point.
+ */
+@Mod("followersloadchunks")
+public class EasyPetsNeoForge {
+    public EasyPetsNeoForge() {
+        EasyPets.init();
+    }
+}
