+++ neoforge/src/main/resources/META-INF/mods.toml
@@ -1,0 +1,7 @@
+modLoader="javafml"
+loaderVersion="[1,)"
+license="MIT"
+[[mods]]
+modId="followersloadchunks"
+version="${file.jarVersion}"
+displayName="EasyPets"
