+++ paper/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    implementation project(":common")
+    compileOnly "io.papermc.paper:paper-api:${paper_version}"
+}
